+++ js/signin.js
@@ -1,3 +1,5 @@
+import { toggleVisiblePassword } from "./sign_module.js";
+
 const emailInput = document.querySelector("#email");
 const passwordInput = document.querySelector("#password");
 const form = document.querySelector("form");
@@ -50,30 +52,11 @@ form.addEventListener("submit", (e) => {
     emailInput.value === "test@codeit.com" &&
     passwordInput.value === "codeit101"
   ) {
-    window.location.href = "./folder";
+    window.location.href = "./folder.html";
   } else {
     focusOutAlert(emailInput, inputErrMessage.emailNotRegistered);
     focusOutAlert(passwordInput, inputErrMessage.passwordNotRegistered);
   }
 });
 
-const visiblePasswordIcon = document.querySelectorAll(".input-wrap > img");
-
-visiblePasswordIcon.forEach((iconEle) => {
-  iconEle.addEventListener("click", () => {
-    const currentType = iconEle.previousElementSibling.getAttribute("type");
-    const inputType = currentType === "password" ? "text" : "password";
-    const imgSrc =
-      currentType === "password"
-        ? "./img/icon/eye_on_icon.svg"
-        : "./img/icon/eye_off_icon.svg";
-    const imgAlt =
-      currentType === "password"
-        ? "현재 비밀번호가 보이는 상태 아이콘"
-        : "현재 비밀번호가 보이지 않는 상태 아이콘";
-
-    iconEle.previousElementSibling.setAttribute("type", inputType);
-    iconEle.setAttribute("src", imgSrc);
-    iconEle.setAttribute("alt", imgAlt);
-  });
-});
+toggleVisiblePassword();
