+++ signin.html
@@ -6,7 +6,7 @@
     <title>로그인</title>
     <link rel="stylesheet" href="./css/reset.css" />
     <link rel="stylesheet" href="./css/sign.css" />
-    <script defer src="./js/signin.js"></script>
+    <script defer type="module" src="./js/signin.js"></script>
   </head>
   <body>
     <main>
