+++ signup.html
@@ -6,7 +6,7 @@
     <title>회원가입</title>
     <link rel="stylesheet" href="./css/reset.css" />
     <link rel="stylesheet" href="./css/sign.css" />
-    <script defer src="./js/signup.js"></script>
+    <script defer type="module" src="./js/signup.js"></script>
   </head>
   <body>
     <main>
